+++ src/main/java/com/elmakers/mine/bukkit/api/block/UndoQueue.java
@@ -1,0 +1,10 @@
+package com.elmakers.mine.bukkit.api.block;
+
+import org.bukkit.block.Block;
+
+public interface UndoQueue {
+	public void add(BlockList blocks);
+	public void removeScheduledCleanup(BlockList blocks);
+	public BlockList getLast();
+	public BlockList getLast(Block target);
+}