+++ src/main/java/com/elmakers/mine/bukkit/api/entity/EntityData.java
@@ -1,0 +1,15 @@
+package com.elmakers.mine.bukkit.api.entity;
+
+import org.bukkit.Art;
+import org.bukkit.Location;
+import org.bukkit.block.BlockFace;
+import org.bukkit.entity.EntityType;
+import org.bukkit.inventory.ItemStack;
+
+public interface EntityData {
+	public Location getLocation();
+	public EntityType getType();
+	public Art getArt();
+	public BlockFace getFacing();
+	public ItemStack getItem();
+}