+++ src/main/java/com/elmakers/mine/bukkit/api/magic/MagicAPI.java
@@ -4,7 +4,6 @@ import java.util.Collection;
 import java.util.logging.Logger;
 
 import org.bukkit.Material;
-import org.bukkit.block.Block;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -401,15 +400,4 @@ public interface MagicAPI {
 	 * @return The list of known schematic names.
 	 */
 	public Collection<String> getSchematicNames();
-	
-	/**
-	 * Check to see if the given block is part of an Automata.
-	 * 
-	 * This currently only works if you specify the redstone block associated
-	 * with the automata, though this will hopefully be improved in the future.
-	 * 
-	 * @param block The block to check.
-	 * @return True if this is an Automata's power source.
-	 */
-	public boolean isAutomata(Block block);
 }